requirements.txt
DVC
numpy 
pandas
matplotlib
seaborne
scikit-learn
plotly